==> profiling/2.A.4_20_Kernel_Weight_Blup_2018-2019.Rmd <==
---
title: "Untitled"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
#knitr::opts_knit$set(root.dir=normalizePath('../../')) 
```



### 2018 DATA
```{r}
#install.packages("readxl")
library(tidyverse)
library(readxl)
ph_2018 <- read.csv("../data/phenotyping_2018_semra_all_traits.csv")
#View(ph_2018)
names(ph_2018)
ph_2018 %>% 
  mutate(pollinationtype = NA, 
         pollinationtype=ifelse(grepl("@",Genotype),"self",pollinationtype),
         pollinationtype=ifelse(grepl("op",tolower(Genotype)),"open",pollinationtype),
         Genotype=gsub("op","",tolower(Genotype)),
         Genotype=gsub("@","",tolower(Genotype)),
         Genotype = str_squish(Genotype)) %>% 
  arrange(Genotype) %>% 
  filter(pollinationtype == "open") %>% 
  rename(genotype = Genotype,
         year = Year) -> ph_2018.b
```


```{r}
ph_2018.b %>% 
  select(genotype, Average.cob.length, Replication, Row) %>% 
  rename(AverageCobLength = `Average.cob.length`) %>% 
  rename(row = `Row`)%>% 
  arrange(genotype) %>% 
  mutate(year = 2018) %>% 
  arrange(genotype) ->  cobl_2018 
  # group_by(genotype) %>% 
  # summarise(meanAvgCob2018 = mean(cobLength2018),
  #           sdAvgCob2018 = sd(cobLength2018))
  
```



```{r}
#View(ph_2018.b)
colnames(ph_2018.b)
ph_2018.b %>%
  rename(row = Row)
colnames(ph_2018.b)[colnames(ph_2018.b) == "Row"] <- "row"
colnames(ph_2018.b)[colnames(ph_2018.b) == "genotype"] <- "Genotype" 
head(ph_2018.b)
unique(ph_2018.b$Genotype)
```


## Burdan duzenle


```{r}
ph_2018.b$Weight.of.20.seeds <- as.numeric(as.character(ph_2018.b$Weight.of.20.seeds))


```

# Merge the CW with the meta data

```{r}
#meta data for 2018
fb <- read.csv("../cache/fb2018_meta.csv")
fb$quadrant <- floor(fb$row/1000)
#View(fb)


#cw$row <- gsub("-.*", "", cw$Line.number) 
#fb1 <- subset(fb, row %% 2 == 1)

#ph_2018.b$Row <- gsub("-.*", "", cw$Line.number) 
fb1 <- subset(fb, row %% 2 == 1)

## remove duplicated rows
idx <- duplicated(ph_2018.b$row)
sub1 <- subset(ph_2018.b, row %in% ph_2018.b[idx,]$row)

cw1 <- subset(ph_2018.b, !(row %in% ph_2018.b[idx,]$row))


fb1cw <- merge(fb1, cw1[, c("row", "Weight.of.20.seeds")], by="row", all.x=TRUE)
#View(fb1cw)

colnames(fb1cw)[colnames(fb1cw) == "pedigree"] <- "Pedigree"
colnames(fb1cw)[colnames(fb1cw) == "genotype"] <- "Genotype"


write.table(fb1cw[, -6], "../cache/pheno2018_20_kernel_weight.csv", sep=",", row.names = FALSE, quote=FALSE)

write_xlsx(fb1cw, "../cache/pheno2018_20_kernel_weight.xlsx")


write.table(subset(fb1cw[, -6], nitrogen %in% "+N"), "../cache/pheno2018_20_kernel_weight_withN.csv", sep=",", row.names = FALSE, quote=FALSE)
write.table(subset(fb1cw[, -6], nitrogen %in% "-N"), "../cache/pheno2018_20_kernel_weight_noN.csv", sep=",", row.names = FALSE, quote=FALSE)

```

# Quality checking

```{r}
# how many of the non-check plots have no data
nonch <- subset(fb1cw, !(Genotype %in% "Check"))
sum(is.na(nonch$Weight.of.20.seeds))

nonch_data <- subset(nonch, !is.na(Weight.of.20.seeds))

t.test(subset(fb1cw, quadrant %in% "1")$Weight.of.20.seeds, subset(fb1cw, quadrant %in% "2")$Weight.of.20.seeds, na.rm=TRUE)

t.test(subset(fb1cw, quadrant %in% "3")$Weight.of.20.seeds, subset(fb1cw, quadrant %in% "4")$Weight.of.20.seeds, na.rm=TRUE)
```

```{r}
library(ggplot2)
fsize=16
p1 <- ggplot(fb1cw, aes(x=Weight.of.20.seeds, color=as.factor(quadrant))) +
      facet_grid(block ~ nitrogen) +
      geom_histogram(bins=30) +
      #labs(y=NULL, fill="Traits") + theme_bw() +
      theme(#axis.text.x=element_blank(), axis.ticks.x=element_blank(),
          axis.text=element_text(size=12),
          axis.title=element_text(size=fsize, face="bold"),
          legend.title = element_text(size=fsize, face="bold"),
          legend.text = element_text(size=fsize),
          strip.text = element_text(size = fsize, face = "bold")) +
      xlab("Weight.of.20.seeds") +
      ylab("Frequency")
p1
```


### 2019 DATA
# Use Cob Width data as an example

```{r}
fb <- read.csv("../cache/fb2019_meta_replaced_check.csv")
fb$quadrant <- floor(fb$row/1000) 
#View(fb)
cw <- read.csv("../data/phenotyping_2019_semra_cobwidth_raw.csv")
table(cw$Number.of.ears)
View(cw)
```


```{r}

cw$Weight.of.20.seeds <- as.numeric(as.character(cw$Weight.of.20.seeds))

```


# Merge the CW with the meta data

```{r}
cw$row <- gsub("-.*", "", cw$Row.number) 
fb1 <- subset(fb, row %% 2 == 1)
#View(fb1)
## remove duplicated rows
idx <- duplicated(cw$row)
sub1 <- subset(cw, row %in% cw[idx,]$row)

cw1 <- subset(cw, !(row %in% cw[idx,]$row))

dim(cw1)
#View(cw1)
fb1cw <- merge(fb1, cw1[, c("row", "Weight.of.20.seeds")], by="row", all.x=TRUE)
#View(fb1cw)
# Quality checking


# how many of the non-check plots have no data
nonch <- subset(fb1cw, !(Genotype %in% "Check"))
sum(is.na(nonch$Weight.of.20.seeds))

nonch_data <- subset(nonch, !is.na(Weight.of.20.seeds))

t.test(subset(fb1cw, quadrant %in% "1")$Weight.of.20.seeds, subset(fb1cw, quadrant %in% "2")$Weight.of.20.seeds, na.rm=TRUE)

t.test(subset(fb1cw, quadrant %in% "3")$Weight.of.20.seeds, subset(fb1cw, quadrant %in% "4")$Weight.of.20.seeds, na.rm=TRUE)


write.table(fb1cw[, -16], "../cache/pheno2019_20_kernel_weight.csv", sep=",", row.names = FALSE, quote=FALSE)

write_xlsx(fb1cw, "../cache/pheno2019_20_kernel_weight.xlsx")


write.table(subset(fb1cw[, -16], nitrogen %in% "+N"), "../cache/pheno2019_20_kernel_weight_withN.csv", sep=",", row.names = FALSE, quote=FALSE)
write.table(subset(fb1cw[, -16], nitrogen %in% "-N"), "../cache/pheno2019_20_kernel_weight_noN.csv", sep=",", row.names = FALSE, quote=FALSE)
```



# Quality checking

```{r}
# how many of the non-check plots have no data
nonch <- subset(fb1cw, !(Genotype %in% "Check"))
sum(is.na(nonch$Weight.of.20.seeds))

nonch_data <- subset(nonch, !is.na(Weight.of.20.seeds))

t.test(subset(fb1cw, quadrant %in% "1")$Weight.of.20.seeds, subset(fb1cw, quadrant %in% "2")$Weight.of.20.seeds, na.rm=TRUE)

t.test(subset(fb1cw, quadrant %in% "3")$Weight.of.20.seeds, subset(fb1cw, quadrant %in% "4")$Weight.of.20.seeds, na.rm=TRUE)
```

```{r}
library(ggplot2)
fsize=16
p2 <- ggplot(fb1cw, aes(x=Weight.of.20.seeds, color=as.factor(quadrant))) +
      facet_grid(block ~ nitrogen) +
      geom_histogram(bins=30) +
      #labs(y=NULL, fill="Traits") + theme_bw() +
      theme(#axis.text.x=element_blank(), axis.ticks.x=element_blank(),
          axis.text=element_text(size=12),
          axis.title=element_text(size=fsize, face="bold"),
          legend.title = element_text(size=fsize, face="bold"),
          legend.text = element_text(size=fsize),
          strip.text = element_text(size = fsize, face = "bold")) +
      xlab("Weight.of.20.seeds") +
      ylab("Frequency")
p2
```


### Merge 2018 and 2019 cob width +N data 
```{r}
### first, we need to organize both data set. (they need to have the same columns)

kernel_weight_2019_N <- read.csv("../cache/pheno2019_20_kernel_weight_withN.csv")

kernel_weight_2018_N <- read.csv("../cache/pheno2018_20_kernel_weight_withN.csv")

#View(coblength_2019_N)
#View(coblength_2018_N)

kernel_weight_18_19_N <- rbind(kernel_weight_2018_N, kernel_weight_2019_N)
#View(cobwidth_18_19)

write.table(kernel_weight_18_19_N, "../cache/20_kernel_weight_2018_2019_N.csv", sep=",", row.names = FALSE, quote=FALSE)
```



### Merge 2018 and 2019 cob width -N data
```{r}
kernel_weight_2018_noN <- read.csv ("../cache/pheno2018_20_kernel_weight_noN.csv")
kernel_weight_2019_noN <- read.csv ("../cache/pheno2019_20_kernel_weight_noN.csv")

kernel_weight_18_19_noN <- rbind (kernel_weight_2018_noN, kernel_weight_2019_noN)
write.table(coblength_18_19_noN, "../cache/20_kernel_weight_2018_2019_noN.csv", sep=",", row.names = FALSE, quote=FALSE)
```


# Install the package

```{r}
#install.packages(devtools)
devtools::install_github("jyanglab/g3tools")
library(g3tools)
```


# Use Cob Width data as an example

```{r}
library(lme4)
#p1 <- read.csv("cache/pheno2019_cw_withN.csv")

get_BLUP(data = kernel_weight_18_19_N , model = Weight.of.20.seeds ~ (1 | Genotype)  + (1 | year) + (1 | block) + (1 | sb) + (1 | spb) + (1| Genotype:year), which.factor = "Genotype",
  outfile = "../cache/BLUP_RESULTS/20_kernel_weight_BLUP_2018_2019_withN_lme4.csv")

#df1 <- read.csv("cache/BLUP/CW_BLUP-2019_withN.csv")
nit <- read.csv("../cache/BLUP_RESULTS/20_kernel_weight_BLUP_2018_2019_withN_lme4.csv")

#cor.test(df1$BLUPs, df2$value)

#plot(df1$BLUPs, df2$value)

#noN
#p2 <- read.csv("cache/pheno2019_cw_noN.csv")

get_BLUP(data = kernel_weight_18_19_noN, model = Weight.of.20.seeds ~ (1 | Genotype) + (1 | year) + (1 | block) + (1 | sb) + (1 | spb) + (1| Genotype:year), which.factor = "Genotype",
  outfile = "../cache/BLUP_RESULTS/20_kernel_weight_BLUP_2018_2019_noN_lme4.csv")

nonit <- read.csv("../cache/BLUP_RESULTS/20_kernel_weight_BLUP_2018_2019_noN_lme4.csv")

#write.table(fb18[, -16:-17], "cache/fb2018_meta.csv", sep=",", row.names = FALSE, quote=FALSE)


#cor.test(df1$value[c(2:215)], df2$value[c(2:215)])

#plot(df1$value[c(2:215)], df2$value[c(2:215)])
```

```{r}

hist(nit$value, breaks=25, col="#6495ed", xlab="20 kernel weight in +N in 2018-2019", ylab="freq", main="histogram")


hist(nonit$value, breaks=25, col="#c6e2ff", xlab="20 kernel weight in -N in 2018-2019", ylab="freq", main="histogram")

t.test(nit$value, nonit$value)


```

## Box Plot
```{r}
library(tidyr)
library(ggplot2)


#setting names
names(nonit) <-c("Genotype","BLUP_noN")
names(nit) <- c("Genotype","BLUP_N")

#checking if data is read in correctly
head(nonit)
head(nit)

#different dimensions show that there are some genotypes in the noN treatment that are in N treatment

dim(nonit)
dim(nit)

#merging the two data sets. all=T means that we include the genotypes that only have 1 data point.
combined<- merge.data.frame(nonit,nit,by="Genotype",all=T)
View(combined)

#convert from wide to long

combinedlong <- gather(combined,Nitrogen,BLUP,c("BLUP_noN","BLUP_N"),factor_key = T)
View(combinedlong)

plot <- ggplot(data=combinedlong,aes(x=Nitrogen,y=BLUP)) + 
        geom_boxplot(outlier.colour="red",notch=T) #red color for outliers, notch=T just means there is a curve in the box plot. Change to notch=F to get a simple box
plot <- plot + scale_x_discrete(labels=c("-N","+N")) + #x-labels
               labs(title="20 Kernel Weight BLUP Boxplots", 
                    xlab="Nitrogen Application",ylab="BLUP") #title and axis labels
plot <- plot + theme(panel.grid.major = element_blank(), 
                     panel.grid.minor = element_blank(),
                     panel.background = element_blank(), 
                     axis.line = element_line(colour = "black"),
                     #legend.text = element_text(size=13,face="bold"), #for legends
                     #legend.title = element_text(size=13, face="bold"), #for legends
                     axis.title.x = element_text(size=18,face="bold"), #format for x-axis
                     axis.title.y = element_text(size=18,face="bold"), #format for y-axis
                     axis.text = element_text(size=18), #format for title
                     plot.margin = unit(c(1,1,1,1),"cm"))#,
                     #legend.position = c(0.8, .75)) #for legend
                      
plot

```




## How to scale all datasets!

```{r}
kwN <- read.csv("../cache/BLUP_RESULTS/20_kernel_weight_BLUP_2018_2019_withN_lme4.csv")
kwnoN <- read.csv("../cache/BLUP_RESULTS/20_kernel_weight_BLUP_2018_2019_noN_lme4.csv")
cwgN <- read.csv("../cache/BLUP_RESULTS/cob_weight_BLUP_2018_2019_withN_lme4.csv")
cwgnoN <- read.csv("../cache/BLUP_RESULTS/cob_weight_BLUP_2018_2019_noN_lme4.csv")
clnN <- read.csv("../cache/BLUP_RESULTS/cob_length_BLUP_2018_2019_withN_lme4.csv")
clnnoN <- read.csv("../cache/BLUP_RESULTS/cob_length_BLUP_2018_2019_noN_lme4.csv")
cwdN <- read.csv("../cache/BLUP_RESULTS/CobWidth_BLUP_2018_2019_noN_lme4.csv")
cwdnoN <- read.csv("../cache/BLUP_RESULTS/CobWidth_BLUP_2018_2019_withN_lme4.csv")
  

#all_blup_data <- cbind (kwN$value, kwnoN$value, cwgN$value, cwgnoN$value, clnN$value, clnnoN$value, cwdN$value, cwdnoN$value)
#write.table(all_blup_data, "../cache/BLUP_RESULTS/all_blup_data.csv", sep=",", row.names = FALSE, quote=FALSE)

all_blup_data <- read.csv("../cache/BLUP_RESULTS/all_blup_data.csv")

summary(all_blup_data)

data_scaled <- as.data.frame(scale(all_blup_data[,c(1:4,6)]))
summary(data_scaled$kwN)
summary(data_scaled$cwgN)

#View(data_scaled)


```




```{r}
install.packages("ggplot2")
library("ggplot2")


#install.packages("tidyverse")
#library("tidyverse")

#devtools::install_github("tidyverse/ggplot2")

View(all_blup_data)

##Prepare the data
all_blup_data$kwN <- as.factor(all_blup_data$kwN)
head(all_blup_data)
str(all_blup_data)

p <- ggplot(all_blup_data, aes(x=kwN, y=)) + 
  geom_violin()
p

ggplot(all_blup_data, aes(x=all_blup_data$kwN, y=all_blup_data$kwnoN)) +
  geom_violin()
# Change the position
p<-ggplot(ToothGrowth, aes(x=dose, y=len, fill=supp)) +
  geom_violin(position=position_dodge(1))
p
```

#VIOLIN PLOT

```{r}
ggplot(data = all_blup_data)

```
