==> profiling/3.A.4_20_Kernel_Weight_Heritability_2018_2019.Rmd <==
---
title: "heritability calculation"
output: html_document
---


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
#knitr::opts_knit$set(root.dir=normalizePath('../../')) 
```



```{r}
library(readr)
kwh <- read.csv("../cache/20_kernel_weight_2018_2019_N.csv")
head(kwh)
class(kwh$Weight.of.20.seeds)
#View(kwh)
head(kwh)
tail(kwh)
dim(kwh)

```



```{r}
install.packages("tidyverse")
update.packages('rlang')
install.packages("devtools")
devtools::install_github("jyanglab/g3tools")
library("g3tools")
```

```{r}
install.packages("scales")
ls(getNamespace("g3tools"), all.names=FALSE)
```


```{r}
options(stringsAsFactors = FALSE) 
install.packages("devtools")
install.packages("Rtools")
library(devtools)
devtools::install_github("jyanglab/g3tools")
devtools::install_github("r-lib/devtools")

```
```{r}
ls(getNamespace("g3tools"), all.names=FALSE)

help("get_BLUP")
help("get_H2")

```



```{r}
pheno <- kwh 


install.packages("Ime4")
library(lme4)
```

```{r}
heritability <- get_BLUP(data = pheno, model = Weight.of.20.seeds ~ (1 | Genotype) + (1 | year) + (1 | block) + (1 | sb) + (1 | spb) + (1
  | Genotype:year)+ (1 | Genotype:block), which.factor = "Genotype", outfile = "heritability.csv")

```


```{r}
summary(heritability)
```


```{r}

Vg= 5.380e-01
Ve= 6.486e-01
Vgxy= 7.265e-02
Vgxb= 2.292e-14
Ny = 2
Nb = 2
NyxNb = 4


H2 = Vg/ (Vg + (Vgxy/Ny) + (Vgxb/Nb) + (Ve/NyxNb))
H2 # 0.7305068

vg = 5.380e-01
ve = 6.486e-01
H2 = vg/(vg+ve/4) #Since I have 2 replications and 2 years. So, it is 2x2= 4
H2 #  0.7684068

```


###Low Nitrogen
```{r}
library(readr)
kwh <- read.csv("../cache/20_kernel_weight_2018_2019_noN.csv")
head(kwh)
class(kwh$Weight.of.20.seeds)
View(kwh)
head(kwh)
tail(kwh)
dim(kwh)

```



```{r}
install.packages("tidyverse")
update.packages('rlang')
install.packages("devtools")
devtools::install_github("jyanglab/g3tools")
library("g3tools")
```

```{r}
install.packages("scales")
ls(getNamespace("g3tools"), all.names=FALSE)
```


```{r}
options(stringsAsFactors = FALSE) 
install.packages("devtools")
install.packages("Rtools")
library(devtools)
devtools::install_github("jyanglab/g3tools")
devtools::install_github("r-lib/devtools")

```
```{r}
ls(getNamespace("g3tools"), all.names=FALSE)

#help("get_BLUP")
#help("get_H2")

```



```{r}
pheno_noN <- kwh 


install.packages("Ime4")
library(lme4)
```

```{r}
heritability <- get_BLUP(data = pheno_noN, model = meancw ~ (1 | Genotype) + (1 | year) + (1 | block) + (1 | sb) + (1 | spb) + (1
  | Genotype:year)+ (1 | Genotype:block), which.factor = "Genotype", outfile = "heritability.csv")

```


```{r}
summary(heritability)
```


```{r}
Vg= 1.481289
Ve= 2.838897
Vgxy= 1.216116
Vgxb= 0
Ny = 2
Nb = 2
NyxNb = 4


H2 = Vg/ (Vg + (Vgxy/Ny) + (Vgxb/Nb) + (Ve/NyxNb))
H2 # 0.5292073


vg = 1.481289
ve = 2.838897
H2 = vg/(vg+(ve/4)) #Since I have 2 replications and 2 years. So, it is 2x2= 4
H2 #  0.6760749

```




```{r}
library(ggplot2)
theme_set(
  theme_minimal() +
    theme(legend.position = "right")
  )

install.packages("ggpubr")
ggpubr::show_point_shapes()



#install.packages("ggplot2")
#install.packages("tidyverse")
library("ggplot2")
heritability <- read.csv("../data/heritability_all.csv")
#View(heritability)

p1n <- ggplot(heritability, aes(x=traits, y=heritability, colour = factor(traits), by = "traits", head = TRUE)) +
    #facet_grid(. ~ type) +
    geom_point(size=12) +
    geom_smooth(method='loess') +
    theme_bw() +
    xlab("Traits") +
    ylab("Heritability") +
    theme(axis.text=element_text(size=14), axis.title=element_text(size=17)) +
  scale_y_continuous(expand = c(0, 0), limits = c(0, 1)) ##force it to start zero
  
#########

p1n
p1n + scale_color_manual(values = c("darkblue", "lightblue", "darkblue", 
                "lightblue", "darkblue", "lightblue", "darkblue", "lightblue")) +
  scale_shape_manual(values = c(3, 23, 18, 16, 17, 19, 20, 21))

  




```

## Bar Chart
```{r}
heritability <- read.csv("../data/heritability_all_copy_copy.csv")
View(heritability)



library(ggplot2)
p <- ggplot(data=heritability, aes(x=Traits, y=Heritability, fill=Treatment)) +
  geom_bar(stat="identity", position=position_dodge())+
  scale_fill_brewer(palette="Paired")+
  theme(axis.text=element_text(size=14), axis.title=element_text(size=17)) 
  #theme_classic()

p + scale_fill_manual(values=c("HN" = "skyblue4","LN"= "skyblue1"))
p + scale_fill_brewer(palette="Blues")

```





```{r}
all_blup<- read.csv("../cache/BLUP_RESULTS/all_blup_data.csv")
View(all_blup)

library(tidyr)
library("ggplot2")
p <- ggplot(c(all_blup$kwN, all_blup$kwnoN, all_blup$cwgN, all_blup$cwgnoN, all_blup$clnN, all_blup$clnnoN, all_blup$cwdN, all_blup$cwdnoN), aes(x = feed, y= weight))


p <- ggplot(all_blup)
p <- geom_boxplot()
p <- theme_classic()
p <- p + labs(title = "Box plot")
p




name <- 



boxplot(all_blup$kwN, all_blup$kwnoN, all_blup$cwgN, all_blup$cwgnoN, all_blup$clnN, all_blup$clnnoN, all_blup$cwdN, all_blup$cwdnoN, main="Radicle Length in salty and nonsalty conditions", col="lightblue", xlab="Traits", ylab="Frequency", names = c("kwN=kernel weight", "clnN=cob length", "cwgN=cob weight", "cwdN=cob width"))
```

```{r}
####SNP-BASED HERITABILITY

snp_heritability<- read.csv("../data/snp_based_heritability.csv")
View(snp_heritability)

library(ggplot2)

snp_heritability_hn_ln<-snp_heritability[c(1,2,6,7,11,12,16,17), ]

p <- ggplot(data=snp_heritability_hn_ln, aes(x=Traits, y=SNP_based_heritability, fill=Treatment)) +
  geom_bar(stat="identity", position=position_dodge())+
  scale_fill_brewer(palette="Paired")+
  theme(axis.text=element_text(size=14), axis.title=element_text(size=17))
  #theme_classic()

p + scale_fill_manual(values=c("HN" = "skyblue4","LN"= "skyblue1"))

 

p <- ggplot(data=snp_heritability, aes(x=Traits, y=SNP_based_heritability, fill=Treatment)) +
  geom_bar(stat="identity", position=position_dodge()) + theme(axis.text=element_text(size=14), axis.title=element_text(size=17)) +
  scale_fill_brewer(direction=-1)
             
 p  

```
